+++ R/group.R
@@ -50,13 +50,13 @@ zarr_group <- R6::R6Class('zarr_group',
       invisible(self)
     },
 
-    #' @description Prints the hierarchy of the group and its subgroups and
-    #'   arrays to the console. Usually called from the Zarr object or its root
-    #'   group to display the full group hierarchy.
+    #' @description Collects the hierarchy of the group and its subgroups and
+    #'   arrays in a character vector. Usually called from the Zarr object or
+    #'   a group to display the full group hierarchy.
     #' @param idx,total Arguments to control indentation. Should both be 1 (the
     #'   default) when called interactively. The values will be updated during
     #'   recursion when there are groups below the current group.
-    hierarchy = function(idx = 1L, total = 1L) {
+    hierarchy_nodes = function(idx = 1L, total = 1L) {
       if (!nzchar(private$.name)) {
         sep <- ''
         knot <- ''
@@ -76,10 +76,18 @@ zarr_group <- R6::R6Class('zarr_group',
       # Sub-groups
       ch <- length(private$.children)
       if (ch > 0L) {
-        sg <- unlist(sapply(1L:ch, function(g) private$.children[[g]]$hierarchy(g, ch)), use.names = FALSE)
+        sg <- unlist(sapply(1L:ch, function(g) private$.children[[g]]$hierarchy_nodes(g, ch)), use.names = FALSE)
         hier <- c(hier, paste0(sep, sg))
       }
       hier
+    },
+
+    #' @description Print the Zarr hierarchy to the console from the current
+    #'   group.
+    hierarchy = function() {
+      cat('<Zarr hierarchy>', self$path, '\n')
+      hier <- self$hierarchy_nodes(1L, 1L)
+      cat(hier, sep = '')
     },
 
     #' @description Return the hierarchy contained in the store as a tree of