+++ NEWS.md
@@ -1,3 +1,5 @@
+# zarr (development version)
+
 # zarr 0.4.0
 
 - Reading of sharded Zarr stores is now supported.