+++ NEWS.md
@@ -1,6 +1,9 @@
 # zarr (development version)
 
-- Hierarchy can now also be printed from any group. Arrays from domain packages may use alternative glyphs.
+- Hierarchy can now also be printed from any group. Zarr arrays from domain packages may use alternative glyphs.
+- New `attribute()` for `zarr_group` and `zarr_array` instances.
+- Attributes can be nested by specifying a compound path when adding. JSON array attributes can be appended. JSON arrays can be deleted over compound paths, including JSON arrays.
+- Attributes article updated.
 
 # zarr 0.4.0