+++ NEWS.md
@@ -1,4 +1,4 @@
-# zarr (development version)
+# zarr 0.4.1
 
 - Hierarchy can now also be printed from any group. Zarr arrays from domain packages may use alternative glyphs.
 - New `attribute()` for `zarr_group` and `zarr_array` instances.